requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=1.5.0
streamlit>=1.36.0
numpy>=1.24.0
datasets>=2.10.0
bitsandbytes>=0.40.0
accelerate>=0.20.0
evaluate>=0.4.0
scikit-learn>=1.2.0
nlpaug>=1.1.0
deep-translator>=1.9.0
psutil>=5.9.0
nltk>=3.8.0

